Move Security settings from Experimental Settings into Security Settings

Move blockaid settings under reveal private key settings.


Fix snapshot.

Fix switch styles
