+++ app/components/Views/Settings/ExperimentalSettings/index.test.tsx
@@ -7,7 +7,6 @@ import { render } from '@testing-library/react-native';
 import initialBackgroundState from '../../../../util/test/initial-background-state.json';
 import { mockTheme, ThemeContext } from '../../../../util/theme';
 import ExperimentalSettings from './';
-import SECURITY_ALERTS_TOGGLE_TEST_ID from './constants';
 
 const mockStore = configureMockStore();
 
@@ -19,36 +18,6 @@ const initialState = {
     backgroundState: initialBackgroundState,
   },
 };
-
-jest.mock('@react-navigation/native', () => ({
-  useNavigation: () => ({
-    navigation: {},
-  }),
-  createNavigatorFactory: () => ({}),
-}));
-
-jest.mock('../../../../util/blockaid', () => ({
-  isBlockaidFeatureEnabled: jest.fn().mockReturnValue(true),
-  isBlockaidSupportedOnCurrentChain: jest.fn().mockReturnValue(true),
-}));
-
-jest.mock('../../../../core/Engine', () => ({
-  context: {
-    PreferencesController: {
-      state: {
-        securityAlertsEnabled: true,
-      },
-      setSecurityAlertsEnabled: () => undefined,
-    },
-    NetworkController: {
-      state: {
-        providerConfig: {
-          chainId: '0x1',
-        },
-      },
-    },
-  },
-}));
 
 const store = mockStore(initialState);
 
@@ -70,28 +39,5 @@ describe('ExperimentalSettings', () => {
       </Provider>,
     );
     expect(wrapper).toMatchSnapshot();
-  });
-
-  it('should render blockaid toggle button', async () => {
-    const wrapper = render(
-      <Provider store={store}>
-        <ThemeContext.Provider value={mockTheme}>
-          <ExperimentalSettings
-            navigation={{
-              setOptions,
-            }}
-            route={{}}
-          />
-          ,
-        </ThemeContext.Provider>
-      </Provider>,
-    );
-    expect(wrapper).toMatchSnapshot();
-    expect(await wrapper.findByText('Security alerts')).toBeDefined();
-    expect(await wrapper.findByText('Blockaid')).toBeDefined();
-
-    const toggle = wrapper.getByTestId(SECURITY_ALERTS_TOGGLE_TEST_ID);
-    expect(toggle).toBeDefined();
-    expect(toggle.props.value).toBe(true);
   });
 });
