+++ app/components/Views/Settings/ExperimentalSettings/index.tsx
@@ -3,7 +3,6 @@ import { Linking, ScrollView, Switch, View } from 'react-native';
 
 import { MMKV } from 'react-native-mmkv';
 import { strings } from '../../../../../locales/i18n';
-import Engine from '../../../../core/Engine';
 import { colors as importedColors } from '../../../../styles/common';
 import { useTheme } from '../../../../util/theme';
 import Text, {
@@ -11,16 +10,10 @@ import Text, {
   TextColor,
 } from '../../../../component-library/components/Texts/Text';
 import { UpdatePPOMInitializationStatus } from '../../../../actions/experimental';
-import AnalyticsV2 from '../../../../util/analyticsV2';
-import { MetaMetricsEvents } from '../../../../core/Analytics';
 import { getNavigationOptionsTitle } from '../../../UI/Navbar';
-import SECURITY_ALERTS_TOGGLE_TEST_ID from './constants';
-import { isBlockaidFeatureEnabled } from '../../../../util/blockaid';
-import Routes from '../../../../constants/navigation/Routes';
-import { useSelector, useDispatch } from 'react-redux';
+import { useDispatch } from 'react-redux';
 import { Props } from './ExperimentalSettings.types';
 import createStyles from './ExperimentalSettings.styles';
-import { selectIsSecurityAlertsEnabled } from '../../../../selectors/preferencesController';
 import Button, {
   ButtonVariants,
   ButtonSize,
@@ -35,11 +28,7 @@ const storage = new MMKV(); // id: mmkv.default
  * Main view for app Experimental Settings
  */
 const ExperimentalSettings = ({ navigation, route }: Props) => {
-  const { PreferencesController } = Engine.context;
-
   const dispatch = useDispatch();
-
-  const securityAlertsEnabled = useSelector(selectIsSecurityAlertsEnabled);
 
   const [sesEnabled, setSesEnabled] = useState(
     storage.getBoolean('is-ses-enabled'),
@@ -55,24 +44,6 @@ const ExperimentalSettings = ({ navigation, route }: Props) => {
   const theme = useTheme();
   const { colors } = theme;
   const styles = createStyles(colors);
-
-  const toggleSecurityAlertsEnabled = () => {
-    if (securityAlertsEnabled) {
-      PreferencesController?.setSecurityAlertsEnabled(false);
-      AnalyticsV2.trackEvent(
-        MetaMetricsEvents.SETTINGS_EXPERIMENTAL_SECURITY_ALERTS_ENABLED,
-        {
-          security_alerts_enabled: false,
-        },
-      );
-    } else if (Device.isIos()) {
-      PreferencesController?.setSecurityAlertsEnabled(true);
-    } else {
-      navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
-        screen: Routes.SHEET.BLOCKAID_INDICATOR,
-      });
-    }
-  };
 
   useEffect(() => {
     dispatch(UpdatePPOMInitializationStatus());
@@ -120,57 +91,6 @@ const ExperimentalSettings = ({ navigation, route }: Props) => {
         width={ButtonWidthTypes.Full}
         style={styles.accessory}
       />
-    </>
-  );
-
-  const BlockaidSettings: FC = () => (
-    <>
-      {Device.isAndroid() && (
-        <Text
-          color={TextColor.Default}
-          variant={TextVariant.HeadingLG}
-          style={styles.heading}
-        >
-          {strings('app_settings.security_heading')}
-        </Text>
-      )}
-      <View style={styles.setting}>
-        <Text color={TextColor.Default} variant={TextVariant.BodyLGMedium}>
-          {strings('experimental_settings.security_alerts')}
-        </Text>
-        <Text
-          color={TextColor.Alternative}
-          variant={TextVariant.BodyMD}
-          style={styles.desc}
-        >
-          {strings('experimental_settings.security_alerts_desc')}
-        </Text>
-      </View>
-      <View style={styles.switchElement}>
-        <Text color={TextColor.Default} variant={TextVariant.BodyLGMedium}>
-          {strings('experimental_settings.blockaid')}
-        </Text>
-        <Switch
-          value={securityAlertsEnabled}
-          onValueChange={toggleSecurityAlertsEnabled}
-          trackColor={{
-            true: colors.primary.default,
-            false: colors.border.muted,
-          }}
-          thumbColor={theme.brandColors.white['000']}
-          style={styles.switch}
-          ios_backgroundColor={colors.border.muted}
-          testID={SECURITY_ALERTS_TOGGLE_TEST_ID}
-        />
-      </View>
-
-      <Text
-        color={TextColor.Alternative}
-        variant={TextVariant.BodyMD}
-        style={styles.desc}
-      >
-        {strings('experimental_settings.blockaid_desc')}
-      </Text>
     </>
   );
 
@@ -222,7 +142,6 @@ const ExperimentalSettings = ({ navigation, route }: Props) => {
     <ScrollView style={styles.wrapper}>
       <WalletConnectSettings />
       {Device.isIos() && <SesSettings />}
-      {isBlockaidFeatureEnabled() && <BlockaidSettings />}
     </ScrollView>
   );
 };
