+++ app/components/Views/Settings/SecuritySettings/Sections/BlockaidSettings.tsx
@@ -1,0 +1,99 @@
+import React from 'react';
+import Device from '../../../../../util/device';
+import Text, {
+  TextVariant,
+  TextColor,
+} from '../../../../../component-library/components/Texts/Text';
+import { useTheme } from '../../../../../util/theme';
+import { strings } from '../../../../../../locales/i18n';
+import createStyles from '../SecuritySettings.styles';
+import { Switch, View } from 'react-native';
+import { useSelector } from 'react-redux';
+import { selectIsSecurityAlertsEnabled } from '../../../../../selectors/preferencesController';
+import Engine from '../../../../../core/Engine';
+import { MetaMetricsEvents } from '../../../../../core/Analytics';
+import AnalyticsV2 from '../../../../../util/analyticsV2';
+import { useNavigation } from '@react-navigation/native';
+import Routes from '../../../../../constants/navigation/Routes';
+import SECURITY_ALERTS_TOGGLE_TEST_ID from '../constants';
+
+const BlockaidSettings = () => {
+  const theme = useTheme();
+  const { colors } = useTheme();
+  const styles = createStyles(colors);
+  const securityAlertsEnabled = useSelector(selectIsSecurityAlertsEnabled);
+  const navigation = useNavigation();
+
+  const toggleSecurityAlertsEnabled = () => {
+    const { PreferencesController } = Engine.context;
+
+    if (securityAlertsEnabled) {
+      PreferencesController?.setSecurityAlertsEnabled(false);
+      AnalyticsV2.trackEvent(
+        MetaMetricsEvents.SETTINGS_EXPERIMENTAL_SECURITY_ALERTS_ENABLED,
+        {
+          security_alerts_enabled: false,
+        },
+      );
+    } else if (Device.isIos()) {
+      PreferencesController?.setSecurityAlertsEnabled(true);
+    } else {
+      navigation.navigate(Routes.MODAL.ROOT_MODAL_FLOW, {
+        screen: Routes.SHEET.BLOCKAID_INDICATOR,
+      });
+    }
+  };
+
+  return (
+    <>
+      {Device.isAndroid() && (
+        <Text
+          color={TextColor.Default}
+          variant={TextVariant.HeadingLG}
+          style={styles.heading}
+        >
+          {strings('app_settings.security_heading')}
+        </Text>
+      )}
+      <View style={styles.setting}>
+        <Text color={TextColor.Default} variant={TextVariant.BodyLGMedium}>
+          {strings('experimental_settings.security_alerts')}
+        </Text>
+        <Text
+          color={TextColor.Alternative}
+          variant={TextVariant.BodyMD}
+          style={styles.desc}
+        >
+          {strings('experimental_settings.security_alerts_desc')}
+        </Text>
+      </View>
+      <View style={styles.blockaidSwitchElement}>
+        <Text color={TextColor.Default} variant={TextVariant.BodyLGMedium}>
+          {strings('experimental_settings.blockaid')}
+        </Text>
+        <Switch
+          value={securityAlertsEnabled}
+          onValueChange={toggleSecurityAlertsEnabled}
+          trackColor={{
+            true: colors.primary.default,
+            false: colors.border.muted,
+          }}
+          thumbColor={theme.brandColors.white['000']}
+          style={styles.switch}
+          ios_backgroundColor={colors.border.muted}
+          testID={SECURITY_ALERTS_TOGGLE_TEST_ID}
+        />
+      </View>
+
+      <Text
+        color={TextColor.Alternative}
+        variant={TextVariant.BodyMD}
+        style={styles.desc}
+      >
+        {strings('experimental_settings.blockaid_desc')}
+      </Text>
+    </>
+  );
+};
+
+export default BlockaidSettings;
